[refactor] add findFeeCurrency util function
