+++ apps/extension/src/hooks/claim/use-cosmos-claim-rewards.ts
@@ -1,11 +1,10 @@
 import { useIntl } from "react-intl";
 import { useStore } from "../../stores";
 import { DefaultGasPriceStep } from "@keplr-wallet/hooks";
-import { CoinPretty, Dec, Int, PricePretty } from "@keplr-wallet/unit";
+import { CoinPretty, Dec, Int } from "@keplr-wallet/unit";
 import {
   AminoSignResponse,
   BroadcastMode,
-  FeeCurrency,
   StdSignDoc,
 } from "@keplr-wallet/types";
 import { InExtensionMessageRequester } from "@keplr-wallet/router-extension";
@@ -20,6 +19,7 @@ import { useNotification } from "../notification";
 import { useNavigate } from "react-router";
 import { NOBLE_CHAIN_ID } from "../../config.ui";
 import { MakeTxResponse } from "@keplr-wallet/stores";
+import { findFeeCurrency } from "../../utils";
 
 const zeroDec = new Dec(0);
 // TODO: Add below property to config.ui.ts
@@ -87,7 +87,6 @@ export const useCosmosClaimRewards = () => {
             (cur) =>
               cur.coinMinimalDenom === chainInfo.stakeCurrency?.coinMinimalDenom
           );
-
       if (chainInfo.hasFeature("osmosis-base-fee-beta") && feeCurrency) {
         const queryBaseFee = queriesStore.get(chainInfo.chainId).osmosis
           .queryBaseFee;
@@ -140,165 +139,12 @@ export const useCosmosClaimRewards = () => {
         }
       }
 
-      if (!feeCurrency) {
-        let prev:
-          | {
-              balance: CoinPretty;
-              price: PricePretty | undefined;
-            }
-          | undefined;
-
-        const feeCurrencies = await (async () => {
-          if (chainInfo.hasFeature("feemarket")) {
-            const queryFeeMarketGasPrices =
-              queriesStore.get(chainId).cosmos.queryFeeMarketGasPrices;
-            await queryFeeMarketGasPrices.waitFreshResponse();
-
-            const result: FeeCurrency[] = [];
-
-            for (const gasPrice of queryFeeMarketGasPrices.gasPrices) {
-              const currency = await chainInfo.findCurrencyAsync(
-                gasPrice.denom
-              );
-              if (currency) {
-                let multiplication = {
-                  low: 1.1,
-                  average: 1.2,
-                  high: 1.3,
-                };
-
-                const multificationConfig = queriesStore.simpleQuery.queryGet<{
-                  [str: string]:
-                    | {
-                        low: number;
-                        average: number;
-                        high: number;
-                      }
-                    | undefined;
-                }>(
-                  "https://gjsttg7mkgtqhjpt3mv5aeuszi0zblbb.lambda-url.us-west-2.on.aws",
-                  "/feemarket/info.json"
-                );
-
-                if (multificationConfig.response) {
-                  const _default =
-                    multificationConfig.response.data["__default__"];
-                  if (
-                    _default &&
-                    _default.low != null &&
-                    typeof _default.low === "number" &&
-                    _default.average != null &&
-                    typeof _default.average === "number" &&
-                    _default.high != null &&
-                    typeof _default.high === "number"
-                  ) {
-                    multiplication = {
-                      low: _default.low,
-                      average: _default.average,
-                      high: _default.high,
-                    };
-                  }
-                  const specific =
-                    multificationConfig.response.data[
-                      chainInfo.chainIdentifier
-                    ];
-                  if (
-                    specific &&
-                    specific.low != null &&
-                    typeof specific.low === "number" &&
-                    specific.average != null &&
-                    typeof specific.average === "number" &&
-                    specific.high != null &&
-                    typeof specific.high === "number"
-                  ) {
-                    multiplication = {
-                      low: specific.low,
-                      average: specific.average,
-                      high: specific.high,
-                    };
-                  }
-                }
-
-                result.push({
-                  ...currency,
-                  gasPriceStep: {
-                    low: parseFloat(
-                      new Dec(multiplication.low)
-                        .mul(gasPrice.amount)
-                        .toString()
-                    ),
-                    average: parseFloat(
-                      new Dec(multiplication.average)
-                        .mul(gasPrice.amount)
-                        .toString()
-                    ),
-                    high: parseFloat(
-                      new Dec(multiplication.high)
-                        .mul(gasPrice.amount)
-                        .toString()
-                    ),
-                  },
-                });
-              }
-            }
-
-            return result;
-          } else {
-            return chainInfo.feeCurrencies;
-          }
-        })();
-        for (const chainFeeCurrency of feeCurrencies) {
-          const currency = await chainInfo.findCurrencyAsync(
-            chainFeeCurrency.coinMinimalDenom
-          );
-          if (currency) {
-            const balance = queries.queryBalances
-              .getQueryBech32Address(account.bech32Address)
-              .getBalance(currency);
-            if (balance && balance.balance.toDec().gt(zeroDec)) {
-              const price = await priceStore.waitCalculatePrice(
-                balance.balance,
-                "usd"
-              );
-
-              if (!prev) {
-                feeCurrency = {
-                  ...chainFeeCurrency,
-                  ...currency,
-                };
-                prev = {
-                  balance: balance.balance,
-                  price,
-                };
-              } else {
-                if (!prev.price) {
-                  if (prev.balance.toDec().lt(balance.balance.toDec())) {
-                    feeCurrency = {
-                      ...chainFeeCurrency,
-                      ...currency,
-                    };
-                    prev = {
-                      balance: balance.balance,
-                      price,
-                    };
-                  }
-                } else if (price) {
-                  if (prev.price.toDec().lt(price.toDec())) {
-                    feeCurrency = {
-                      ...chainFeeCurrency,
-                      ...currency,
-                    };
-                    prev = {
-                      balance: balance.balance,
-                      price,
-                    };
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
+      feeCurrency = await findFeeCurrency({
+        chainInfo,
+        queriesStore,
+        priceStore,
+        account,
+      });
 
       if (feeCurrency) {
         try {
