+++ apps/extension/src/utils/index.ts
@@ -4,3 +4,4 @@ export * from "./amount-to-ambiguous-string";
 export * from "./side-panel";
 export * from "./set-interaction-data-href";
 export * from "./hex-to-rgba";
+export * from "./fine-fee-currency";
